% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzylink_with_embeddings.R
\name{fuzzylink_with_embeddings}
\alias{fuzzylink_with_embeddings}
\title{Probabilistic Record Linkage Using Pretrained Text Embeddings with embeddings already found
This code is almost identical to the fuzzylink code, but does not get the embeddings--these are passed as a parameter}
\usage{
fuzzylink_with_embeddings(
  dfA,
  dfB,
  embeddings,
  by,
  blocking.variables = NULL,
  verbose = TRUE,
  record_type = "entity",
  instructions = NULL,
  model = "gpt-4o-2024-11-20",
  openai_api_key = Sys.getenv("OPENAI_API_KEY"),
  learner = "glm",
  fmla = match ~ sim + jw,
  max_labels = 10000,
  parallel = TRUE,
  return_all_pairs = FALSE,
  text_gen_port_num = 8081,
  debug = FALSE
)
}
\arguments{
\item{dfA, dfB}{A pair of data frames or data frame extensions (e.g. tibbles)}

\item{embeddings}{The embeddings of dfA and dfB}

\item{by}{A character denoting the name of the variable to use for fuzzy matching}

\item{blocking.variables}{A character vector of variables that must match exactly in order to match two records}

\item{verbose}{TRUE to print progress updates, FALSE for no output}

\item{record_type}{A character describing what type of entity the \code{by} variable represents. Should be a singular noun (e.g. "person", "organization", "interest group", "city").}

\item{instructions}{A string containing additional instructions to include in the LLM prompt during validation.}

\item{model}{Which LLM to prompt when validating matches; defaults to 'gpt-4o-2024-11-20    '. Set to "EMPTY" to run locally.}

\item{openai_api_key}{Your OpenAI API key. By default, looks for a system environment variable called "OPENAI_API_KEY" (recommended option). Otherwise, it will prompt you to enter the API key as an argument.}

\item{learner}{Which supervised learner should be used to predict match probabilities. Defaults to logistic regression ('glm'), but will also accept random forest ('ranger').}

\item{fmla}{By default, logistic regression model predicts whether two records match as a linear combination of embedding similarity and Jaro-Winkler similarity (\code{match ~ sim + jw}). Change this input for alternate specifications.}

\item{max_labels}{The maximum number of LLM prompts to submit when labeling record pairs. Defaults to 10,000}

\item{parallel}{TRUE to submit API requests in parallel. Setting to FALSE can reduce rate limit errors at the expense of longer runtime.}

\item{return_all_pairs}{If TRUE, returns \emph{every} within-block record pair from dfA and dfB, not just validated pairs. Defaults to FALSE.}

\item{text_gen_port_num}{The port number that the local text generation model is running on. Defaults to 8081.}

\item{debug}{TRUE to print various statments throughout the code to track progess. Defaults to FALSE.}
}
\value{
A dataframe with all rows of \code{dfA} joined with any matches from \code{dfB}
}
\description{
Probabilistic Record Linkage Using Pretrained Text Embeddings with embeddings already found
This code is almost identical to the fuzzylink code, but does not get the embeddings--these are passed as a parameter
}
