% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_embeddings.R
\name{get_embeddings}
\alias{get_embeddings}
\title{Get pretrained text embeddings}
\usage{
get_embeddings(
  text,
  model = "text-embedding-3-large",
  dimensions = 256,
  openai_api_key = Sys.getenv("OPENAI_API_KEY"),
  parallel = TRUE,
  port_number = 8080,
  debug = FALSE
)
}
\arguments{
\item{text}{A character vector}

\item{model}{Which embedding model to use. Defaults to 'text-embedding-3-large'. Set to "EMPTY" to run locally.}

\item{dimensions}{The dimension of the embedding vectors to return. Defaults to 256. Note that the 'mistral-embed' model will always return 1024 vectors.}

\item{openai_api_key}{Your OpenAI API key. By default, looks for a system environment variable called "OPENAI_API_KEY".}

\item{parallel}{TRUE to submit API requests in parallel. Setting to FALSE can reduce rate limit errors at the expense of longer runtime.}

\item{port_number}{The port number that the local embedding model is running on. Defaults to 8080.}

\item{debug}{TRUE to print various statments throughout the code to track progess. Defaults to FALSE.}
}
\value{
A matrix of embedding vectors (one per row).
}
\description{
Get pretrained text embeddings from the OpenAI or Mistral API. Automatically batches requests to handle rate limits.
}
\examples{
\dontrun{
embeddings <- get_embeddings(c('dog', 'cat', 'canine', 'feline'))
embeddings['dog',] |> dot(embeddings['canine',])
embeddings['dog',] |> dot(embeddings['feline',])
}
}
